backend/requirements.txt
runpod
boto3
accelerate==1.8.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==25.3.0
audioread==3.0.1
av==15.0.0
Brotli==1.1.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comfyui-embedded-docs==0.2.4
comfyui_frontend_package==1.23.4
comfyui_workflow_templates==0.1.35
decorator==5.2.1
demucs==4.0.1
diffusers==0.34.0
dora_search==0.1.12
einops==0.8.1
exceptiongroup==1.3.0
fastapi==0.116.1
ffmpy==0.6.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.3.1
future==1.0.0
gguf==0.17.1
glcontext==3.0.0
gradio==5.36.2
gradio_client==1.10.4
greenlet==3.2.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.1
julius==0.2.7
kornia==0.8.1
kornia_rs==0.1.9
lameenc==1.8.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moderngl==5.12.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
networkx==3.4.2
noise==1.2.2
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
opencv-python==4.12.0.88
openunmix==1.3.0
orjson==3.10.18
packaging==25.0
pandas==2.3.1
peft
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pygame==2.6.1
Pygments==2.19.2
pyloudnorm==0.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
retrying==1.4.0
rich==14.0.0
ruff==0.12.3
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
spandrel==0.4.1
SQLAlchemy==2.0.41
starlette==0.47.1
submitit==1.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
tomlkit==0.13.3
torch==2.7.1
torchaudio==2.7.1
torchsde==0.2.6
torchvision==0.22.1
tqdm==4.67.1
trampoline==0.1.2
transformers==4.53.2
treetable==0.2.5
triton==3.3.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0

